feat: switched the keyboard shortcuts
Switched keyboard shortcuts for jumping to tab n on the results list and toggling mute for tab n on
the list as I feel that user would mostly use the arrow keys and enter to do that. Another reason is
the keyboard interaction to toggle mute for a given tab is too much work, this should make it much
easier.