feat: show recently audible tabs in the show audibly only view
The show audible tabs only filtered view will now also start showing the tabs which aren't audible
right now but were audible in the last t milliseconds. t is set to 5 minutes for now, but will
probably be made user-configurable in the future. The recently audible tabs list is sorted with the
most recently audible at the top of the list.